+++ MavenTienda/src/test/java/almacen/Test_Productos.java
@@ -1,0 +1,44 @@
+package almacen;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class Test_Productos {
+
+	@Test
+	void testProductos() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testGetNombre() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testSetNombre() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testGetPrecio() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testSetPrecio() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testGetCantidad() {
+		fail("Not yet implemented");
+	}
+
+	@Test
+	void testSetCantidad() {
+		fail("Not yet implemented");
+	}
+
+}
